Extract version comparison to separate module (with tests)
I wanted to re use version sorting in several places so it seemed
logical to group the related code together in a new `version' module.

While I was at it I decided to write tests that make sure the results of
compare_versions_with_python_apt() and compare_versions_with_dpkg()
are consistent with each other and the expected behavior.