Promote python-debian constraint to conditional dependency*
Recently I constrained the version of python-debian to work around a
Python 2.6 incompatibility in 24a5a16ef02. This same incompatibility is
now biting me in the py2deb setup on Travis CI [1] and after fighting
that situation for a while I decided it may be better / less convoluted
to fix this in deb-pkg-tools instead (at the source of the problem, so
to speak).

[1] https://github.com/paylogic/py2deb/compare/4284a1db99699bab14bc5fb62a88256a5d1ae978...60ece9ffebbd5f1bdff7ea20fbf0eeb401a9da3f