Attempt to stabilize Travis CI builds ...
The previous commit (a21a7358aa8) was intended to expose a specific test
failure which it didn't, however it did expose other build failures:

  https://travis-ci.org/xolox/python-deb-pkg-tools/builds/481390099

I suspect at least some of these may be related to the old version of
pip that pip-accel depends on. In order to confirm that suspicion I'm
removing the usage of pip-accel from the makefile, to see if constraints
are respected in a more sane way (not invalidating constraints of
installed packages).