Simplify overview of environment variables
Over the years I've picked up the habit of treating my documentation
just like my code: Make sure everything is defined in a single place
(DRY), as close as possible to the place where it is used.

Properly documenting all of the module variables that are based on
environment variables and linking to those from the readme has freed me
from the burden of explaining things in more than one place.

This is good because multiple explanations increase the chance of
documentation becoming outdated which is definitely something to be
avoided whenever possible.