Start using __all__ to control what is exported
This change is backwards incompatible in the sense that until now
imports were exposed to the outside world, however for anyone to
actually use this would imply not having read the documentation,
so this doesn't really bother me.

In theory this change could be backwards incompatible in a bad if I
omitted __all__ entries that should have been exported. I did double
check but of course I can't be 100% sure (the deb_pkg_tools.* modules
currently span almost 6000 lines including whitespace and comments).